genaiDocumentSummarizer/backend/requirements.txt
flask 
flask-cors 
python-dotenv 
langchain 
langchain-community 
PyPDF2 
openai 
faiss-cpu
